Move UrbanEventType tests to EventConfig tests in imio.urban.core.
