+++ src/Products/urban/tests/test_UrbanTemplate.py
@@ -4,6 +4,7 @@ from Products.urban.interfaces import IEnvironmentBase
 from Products.urban.profiles.testsWithLicences.licences_data import licences_data
 from Products.urban.testing import URBAN_TESTS_LICENCES
 
+from plone import api
 from plone.app.testing import login
 
 import unittest
@@ -61,3 +62,24 @@ class TestTemplateMethods(unittest.TestCase):
                 template_helpermethod = getattr(licence, method_name, None)
                 if template_helpermethod:
                     template_helpermethod()
+
+    def testUrbanTemplateIsUnderActivationWF(self):
+        wf_tool = api.portal.get_tool('portal_workflow')
+        # Check that templates .odt files in urbanEventTypes are under activation wf policy
+        urban_event_type = getattr(self.portal_urban.buildlicence.eventconfigs, 'accuse-de-reception', None)
+        template = getattr(urban_event_type, 'urb-accuse.odt', None)
+        state = wf_tool.getInfoFor(template, 'review_state')
+        self.assertEqual(state, 'enabled')
+
+    def testGeneratedDocumentIsNotUnderActivationWF(self):
+        wf_tool = api.portal.get_tool('portal_workflow')
+
+        # Check that generated .odt files in urbanEvents are NOT under any wf policy
+        urban_event = self.buildlicence.getLastAcknowledgment()
+        document = getattr(urban_event, 'urb-accuse.odt', None)
+        exception_msg = ""
+        try:
+            wf_tool.getInfoFor(document, 'review_state')
+        except Exception, error:
+            exception_msg = "%s" % error
+        self.assertEqual(exception_msg, "No workflow provides '${name}' information.")
